Fix admin translations not being loaded in Safari
The t() function fallback which was added in #1811 seems to be declared
even if the surrounding if() block does not resolve to true, resulting
in t() always being the fallback method instead of the real translator.
By setting it as function expression on the window object the fallback
should work properly without overwriting the actual t() function.

Fixes #1997